gặp lỗi initial_value must be str or None, not bytes

+++ dga_classifier/dga_generators/dircrypt.py
@@ -8,7 +8,7 @@ class RandInt:
 
     def rand_int_modulus(self, modulus):
         ix = self.seed                
-        ix = 16807*(ix % 127773) - 2836*(ix / 127773) & 0xFFFFFFFF        
+        ix = 16807*int(ix % 127773) - 2836*int(ix / 127773) & 0xFFFFFFFF        
         self.seed = ix
         return ix % modulus
